move navbar to tapis-ui
