+++ src/tapis-ui/components/Navbar/Navbar.module.scss
@@ -1,6 +1,6 @@
 $core-portal-color-primary: #9d85ef;
 
-.apps-list {
+.nav-list {
   --horizontal-buffer: var(--global-space--section-left);
   --vertical-buffer: 0.75rem;
   --min-width: 100%;
