+++ src/tapis-ui/components/apps/AppsListing/AppsListing.tsx
@@ -3,7 +3,7 @@ import { useRouteMatch } from 'react-router-dom';
 import { useList } from 'tapis-hooks/apps';
 import { LoadingSpinner, Message } from 'tapis-ui/_common';
 import { Apps } from '@tapis/tapis-typescript';
-import { Navbar, NavItem } from 'tapis-app/Navbar';
+import { Navbar, NavItem } from 'tapis-ui/components/Navbar';
 
 const AppsListing: React.FC = () => {
   const { data, isLoading, error } = useList();
